Initial support for unlimited memory region overlays
Previously the way the memory region was resolved meant that there
could only be one region overlayed on top of the main RAM region.

This commit updates the algorithm used to insert and resolve regions
to support theoretically unlimited memory region overlays, but
performance suffers the more overlays there are.