cpu: arithmetic: Fix a couple of arithmetic issues
First, result_size is required for AND operations but we so far have
not been affected by it.

Second, the order of operations for inc was wrong causing the result
to be 0 when it's stored. This led to jnz never working correctly.

Finally, the check for ZF was wrong. It returned true if it wasn't
zero but false if it actually was zero.